API

LatticeUtilities.Bond

LatticeUtilities.Bond

LatticeUtilities.Lattice

LatticeUtilities.Lattice

LatticeUtilities.UnitCell

LatticeUtilities.UnitCell

LatticeUtilities.UnitCell

LatticeUtilities.bond_to_vec

LatticeUtilities.bond_to_vec!

LatticeUtilities.build_neighbor_table

LatticeUtilities.build_neighbor_table

LatticeUtilities.calc_k_point

LatticeUtilities.calc_k_point!

LatticeUtilities.calc_k_points

LatticeUtilities.calc_k_points!

LatticeUtilities.displacement_to_vec

LatticeUtilities.displacement_to_vec!

LatticeUtilities.get_num_sites

LatticeUtilities.loc_to_pos

LatticeUtilities.loc_to_pos

LatticeUtilities.loc_to_pos!

LatticeUtilities.loc_to_pos!

LatticeUtilities.loc_to_site

LatticeUtilities.loc_to_site

LatticeUtilities.loc_to_unitcell

LatticeUtilities.map_neighbor_table

LatticeUtilities.pbc!

LatticeUtilities.simplify!

LatticeUtilities.simplify!

LatticeUtilities.site_to_loc

LatticeUtilities.site_to_loc!

LatticeUtilities.site_to_orbital

LatticeUtilities.site_to_site

LatticeUtilities.site_to_unitcell

LatticeUtilities.sites_to_bond

LatticeUtilities.sites_to_bond!

LatticeUtilities.sites_to_displacement

LatticeUtilities.sites_to_displacement!

LatticeUtilities.translational_avg!

LatticeUtilities.unitcell_to_loc

LatticeUtilities.unitcell_to_loc!

LatticeUtilities.valid_loc

LatticeUtilities.valid_site

LatticeUtilities.Bond — Type

Bond

Defines a bond defintion in lattice.

Fields

D::Int64

Dimension of system bond is in.

orbitals::Vector{Int64}

Initial and final orbital species respectively.

displacement::Vector{Int64}

Displacement in unit cells.

LatticeUtilities.Bond — Method

Bond(orbitals::AbstractVector{Int}, displacement::AbstractVector{Int})

Constrcut a Bond

LatticeUtilities.Lattice — Type

Lattice

A type defining a finite lattice in arbitary dimensions.

Fields

D::Int64

Number of spatial dimensions.

N::Int64

Number of unit cells.

L::Vector{Int64}

Linear extent of lattice in the direction of each lattice vector.

periodic::Vector{Bool}

Whether the lattice is periodic in the direction of each lattice vector.

lvec::Vector{Int64}

Storage space for representing a location or displacement in the lattice.

LatticeUtilities.Lattice — Method

Lattice(L::Vector{Int}, periodic::Vector{Bool})

Constructs a Lattice.

LatticeUtilities.UnitCell — Type

UnitCell{T<:AbstractFloat}

A type defining a unit cell.

Fields

D::Int64

Number of spatial dimensions.

n::Int64

Orbitals per unit cell.

lattice_vecs::Matrix{T} where T<:AbstractFloat

Matrix where the columns are the lattice vectors.

reciprocal_vecs::Matrix{T} where T<:AbstractFloat

Matrix where the columns are the reciprocal lattice vectors.

basis_vecs::Matrix{T} where T<:AbstractFloat

Matrix where the columns are then basis vectors.

LatticeUtilities.UnitCell — Method

UnitCell(lattice_vecs::AbstractVector{Vector{T}}, basis_vecs::AbstractVector{Vector{T}})
    where {T<:AbstractFloat}

Constrcuts a UnitCell.

LatticeUtilities.UnitCell — Method

UnitCell(lattice_vecs::Matrix{T}, basis_vecs::Matrix{T}) where {T<:AbstractFloat}

Constrcuts a UnitCell.

Base.:== — Method

Base.:==(b₁::Bond,b₂::Bond)

Tests if two bonds b₁ and b₂ are equivalent.

Base.show — Method

Base.show(io::IO, bond::Bond)
Base.show(io::IO, ::MIME"text/plain", bond::Bond)

Show lattice.

Base.show — Method

Base.show(io::IO, lattice::Lattice)
Base.show(io::IO, ::MIME"text/plain", lattice::Lattice)

Show lattice.

Base.show — Method

Base.show(io::IO, uc::UnitCell{T}) where {T}
Base.show(io::IO, ::MIME"text/plain", uc::UnitCell{T}) where {T}

Show unit cell.

LatticeUtilities.bond_to_vec! — Method

bond_to_vec!(Δr::AbstractVector{T}, bond::Bond, unit_cell::UnitCell{T}) where {T}

Calculate the displacement vector associated with a bond.

LatticeUtilities.bond_to_vec — Method

bond_to_vec(bond::Bond, unit_cell::UnitCell{T})

Return the displacement vector associated with a bond.

LatticeUtilities.build_neighbor_table — Method

build_neighbor_table(bonds::AbstractVector{Bond}, unit_cell::UnitCell, lattice::Lattice)

Construct the neighbor table corresponding to bonds.

LatticeUtilities.build_neighbor_table — Method

build_neighbor_table(bond::Bond, unit_cell::UnitCell, lattice::Lattice)

Construct the neighbor table corresponding to bond.

LatticeUtilities.calc_k_point! — Method

calc_k_point!(k_point::AbstractVector{T}, k_loc::AbstractVector{Int},
    unit_cell::UnitCell{T}, lattice::Lattice) where {T}

Calculate the k-point k_point corresponding to the k-point location k_loc.

LatticeUtilities.calc_k_point — Method

calc_k_point(k_point::AbstractVector{T}, k_loc::AbstractVector{Int},
    unit_cell::UnitCell{T}, lattice::Lattice) where {T}

Return the k-point k_point corresponding to the k-point location k_loc.

LatticeUtilities.calc_k_points! — Method

calc_k_points!(k_points::AbstractArray{T}, unit_cell::UnitCell{T},
    lattice::Lattice) where {T}

Calculate the k-point grid k_points assicated with a finite lattice.

LatticeUtilities.calc_k_points — Method

calc_k_points!(unit_cell::UnitCell{T}, lattice::Lattice) where {T}

Return the k-point grid assicated with a finite lattice. For a D dimensional lattice, a D+1 dimensional array will be returned. If the system has open boundary conditions in a given direction, it will treat the linear extent of the system in that direction as equalling L=1 for the purposes of calucting the k-points.

LatticeUtilities.displacement_to_vec! — Method

displacement_to_vec!(Δr::AbstractVector{T}, Δl::AbstractVector{Int},
    o₁::Int, o₂::Int, unit_cell::UnitCell{T}) where {T}

Computes the position space displacement vector Δr corresponding to a displacement definition given by initial and final orbitals o₁ and o₂ in the unit cell respectively, along with a displacement in unit cells Δl.

Arguments

Δr::AbstractVector{T}: displacement vector in position space.

Δl::AbstractVector{Int}: displacement in unit cells.

o₁::Int: initial orbital in unit cell.

o₂::Int: final orbital in unit cell.

unit_cell::UnitCell{T}: unit cell.

LatticeUtilities.displacement_to_vec — Method

displacement_to_vec(Δl::AbstractVector{Int}, o₁::Int, o₂::Int,
    unit_cell::UnitCell{T}) where {T}

Returns the position space displacement vector Δr corresponding to a displacement definition given by initial and final orbitals o₁ and o₂ in the unit cell respectively, along with a displacement in unit cells Δl.

LatticeUtilities.get_num_sites — Method

get_num_sites(unit_cell::UnitCell, lattice::Lattice)

Returns the number of sites Nₛ in a finite lattice.

LatticeUtilities.loc_to_pos! — Method

loc_to_pos!(r::AbstractVector{T}, l::AbstractVector{Int}, s::Int,
    unit_cell::UnitCell{T}) where {T}

Calculate the position r of a orbital o at location l.

LatticeUtilities.loc_to_pos! — Method

loc_to_pos!(r::AbstractVector{T}, l::AbstractVector{Int},
    unit_cell::UnitCell{T}) where {T}

Calculate the position r of a unit cell at location l.

LatticeUtilities.loc_to_pos — Method

loc_to_pos(l::AbstractVector{Int}, s::Int, unit_cell::UnitCell{T})
    where {T}

Return the position r of a orbital o at location l.

LatticeUtilities.loc_to_pos — Method

loc_to_pos(l::AbstractVector{Int}, unit_cell::UnitCell{T})
    where {T}

Return the position r of a unit cell at location l.

LatticeUtilities.loc_to_site — Method

loc_to_site(l::AbstractVector{Int}, o::Int, unit_cell::UnitCell, lattice::Lattice)

Given a unit cell location l and orbital species o, return the corresponding site s in the lattice. If the location is not valid owing to open boundary conditions then return s = 0.

LatticeUtilities.loc_to_site — Method

loc_to_site(u::Int, o::Int, unit_cell::UnitCell, lattice::Lattice)

Given a unit cell index u and orbital o, return the correspond site s.

LatticeUtilities.loc_to_unitcell — Method

loca_to_unitcell(l::AbstractVector{Int}, lattice::Lattice)

Return the unit cell found at location l in the lattice.

LatticeUtilities.map_neighbor_table — Method

map_neighbor_table(neighbor_table::Matrix{Int})

For a given neighbor table, return a dictionary that reports the bonds and neighbors associated with each site in the lattice. If neighbor_table is modified, then a new map must be constructed.

LatticeUtilities.pbc! — Method

pbc!(l::AbstractVector{Int}, lattice::Lattice)

Apply periodic boundary to unit cell location l.

LatticeUtilities.simplify! — Method

simplify!(Δl::AbstractVector{Int}, lattice::Lattice)

Simplify displacement Δl so that it is as short as possible accounting for periodic boundary conditions where necessary.

LatticeUtilities.simplify! — Method

simplify!(bond::Bond, lattice::Lattice)

Simplify a bond so that the displacement is the shortest possible accounting for periodic boundary conditions where necessary.

LatticeUtilities.site_to_loc! — Method

site_to_loc!(l::AbstractVector{Int}, s::Int, unit_cell::UnitCell, lattice::Lattice)

For a given site s in the lattice, calculate the location l of the unit cell it is in and return the orbital species o of the site.

LatticeUtilities.site_to_loc — Method

site_to_loc(s::Int, unit_cell::UnitCell, lattice::Lattice)

For a given site s in the lattice, return the location l of the unit cell it is in and the orbital species o.

LatticeUtilities.site_to_orbital — Method

site_to_orbital(s::Int, unit_cell::UnitCell, lattice::Lattice)

Return the orbtial species of site s.

LatticeUtilities.site_to_site — Method

site_to_site(s₁::Int, Δl::AbstractVector{Int}, o₂::Int, unit_cell::UnitCell, lattice::Lattice)

Given an initial site s₁, and a displacement in unit cells Δl and a terminating orbital species o₂, return the resulting site s₂ in the lattice. If the displacement is not allowed as a result of open boundary conditions, then s₂=0 is returned.

LatticeUtilities.site_to_unitcell — Method

site_to_unitcell(s::Int, unit_cell::UnitCell, lattice::Lattice)

Return the unit cell u containing lattice site s.

LatticeUtilities.sites_to_bond! — Method

sites_to_bond!(bond::Bond, s₁::Int, s₂::Int, unit_cell::UnitCell, lattice::Lattice)

Calculate the bond associated with getting displaced from site s₁ to s₂.

LatticeUtilities.sites_to_bond — Method

sites_to_bond(s₁::Int, s₂::Int, unit_cell::UnitCell, lattice::Lattice)

Return the bond associated with getting displaced from site s₁ to s₂.

LatticeUtilities.sites_to_displacement! — Method

function sites_to_displacement!(o::AbstractVector{Int}, Δl::AbstractVector{Int},
    s₁::Int, s₂::Int, unit_cell::UnitCell, lattice::Lattice)

When getting displaced from site s₁ to s₂, calculate the initial and final orbitals o and displacement in unit cells Δl.

LatticeUtilities.sites_to_displacement — Method

function sites_to_displacement(s₁::Int, s₂::Int, unit_cell::UnitCell, lattice::Lattice)

When getting displaced from site s₁ to s₂, return the initial and final orbitals o and displacement in unit cells Δl.

LatticeUtilities.translational_avg! — Method

function translational_avg!(fg::AbstractArray{Complex{T}}, f::AbstractArray{Complex{T}},
    g::AbstractArray{Complex{T}}; restore::Bool=true) where {T<:AbstractFloat}

Let f[i] and g[j] be two distinct multi-dimensional arrays, where i and j represent an index into them and also correspond to the position of a unit cell in a periodic finite lattice This method then computes in-place the product (f⋅g)[i-j] that is averaged over translation symmetry. If restore = true then f and g are left unchanged, otherwise they will be left modified.

LatticeUtilities.unitcell_to_loc! — Method

unitcell_to_loc!(l::AbstractVector{Int}, u::Int, lattice::Lattice)

Calculate the location l of a unit cell u.

LatticeUtilities.unitcell_to_loc — Method

unitcell_to_loc(u::Int, lattice::Lattice)

Return the location of unit cell u.

LatticeUtilities.valid_loc — Method

valid_location(l::AbstractVector{Int}, lattice::Lattice)

Determine if l describes a valid location in the lattice.

LatticeUtilities.valid_site — Method

valid_site(s::Int, unit_cell::UnitCell, lattice::Lattice)

Return whether s is a valid site index.